Yoni ios ui merge 2 android (#36)
* ios ui android merge
specialBle start scan adjustment

* Home android restored to original
UI fixes (safeAreaView)

* IOS home screen ui fixes
Data model fixes
Results screeen data fixes
advertisementData data conversion for dict

* revert ios signing configuration to be manual

* results null public_key fix
specialBLE scan device properties types buildup fixes

* fixed a bit the UI

* Home screen ui fixes - durations/intervals
Config fixes
GetConfig fix
* if empty or null value will get, the default value will be assigned

* Home ios missing devices fix
Results crash fix
SpecialBle CBPeripheralDelegate conformation added
SpecialBleManager publicKey- fix

* automatic signing
StartBleServices
start / stop services flow fix
nil publicKey advertise bug fix
timeStamps properties fix (int64)
addDevice / updateDevice functionality
DBManager fetchDeviceByKey fix
DBManager updateDevice implementation
DB models fixes

* Home iOS ui fixes: alerts removed, Config section hidden, cleanAllScans added
Results UI fix: minutes and seconds are 2 digits
Scans UI fix: minutes and seconds are 2 digits
SpecialBle scans method fix
SpecialBleManager : start/stop services full flow fix, start/stop tasks full flow fix, scans added to DB by central
DB Scans methods added
DB Scans functionality fix
Scan model fixes

* Logs for unKnown device added

+++ lib/ios/db/CoreData/Device+CoreDataProperties.swift
@@ -21,8 +21,8 @@ extension Device {
     @NSManaged public var device_address: String?
     @NSManaged public var device_protocol: String?
     @NSManaged public var device_rssi: Int16
-    @NSManaged public var device_first_timestamp: Int16
-    @NSManaged public var device_last_timestamp: Int16
+    @NSManaged public var device_first_timestamp: Int64
+    @NSManaged public var device_last_timestamp: Int64
     @NSManaged public var device_tx: Int16
 
 }
